+++ src/tree/items/NoSolutionTreeItem.ts
@@ -4,7 +4,7 @@ import { TreeItemContext } from "../TreeItemContext";
 
 export class NoSolutionTreeItem extends TreeItem {
     constructor(context: TreeItemContext, rootPath: string) {
-        super(context, 'No solution found (click to create a new one)', TreeItemCollapsibleState.None, ContextValues.NoSolution, rootPath);
+        super(context, 'No solution found (right click to create a new one)', TreeItemCollapsibleState.None, ContextValues.NoSolution, rootPath);
         this.iconPath = null;
     }
 }