+++ src/commands/CleanCommand.ts
@@ -11,12 +11,10 @@ export class CleanCommand extends CliCommandBase {
 
     protected shouldRun(item: TreeItem): boolean {
         this.parameters = [
-            new StaticCommandParameter('clean')
+            new StaticCommandParameter('clean'),
+            new StaticCommandParameter(item.path)
         ];
-
-        if (item.contextValue.startsWith(ContextValues.Project))
-            this.parameters.push(new StaticCommandParameter(item.path));
-
+        
         return true;
     }
 }