+++ src/commands/RunCommand.ts
@@ -12,7 +12,7 @@ export class RunCommand extends CliCommandBase {
     protected shouldRun(item: TreeItem): boolean {
         this.parameters = [
             new StaticCommandParameter('run'),
-            new StaticCommandParameter(item.path)
+            new StaticCommandParameter(item.path, '--project')
         ];
 
         return true;