+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Bug fixed #51: Updated license to CC-BY-SA.
 
 Bug fixed #49: Modify default 'showMode' and 'trackActiveItem' configuration setting values.
 
+New comand parameter compiler adding a wizard style requesting commad parameters.
+
+Bug fixed #48: now you can specify the package version when adding a new nuget package to a project.
+
 ## 0.2.34
 
 Added "win32Encoding" configuration parameter to help people fixing win32 encoding problems in the output.