+++ CHANGELOG.md
@@ -12,6 +12,8 @@ New comand parameter compiler adding a wizard style requesting commad parameters
 
 Bug fixed #48: now you can specify the package version when adding a new nuget package to a project.
 
+Bug fixed #55: changed message to create a new solution.
+
 ## 0.2.34
 
 Added "win32Encoding" configuration parameter to help people fixing win32 encoding problems in the output.